Clean more cases of 'integer for DB rowid' values (bgo#696520)
The transition from int to long types [1] had some leftovers
which were working as runtime casts.

[1] https://git.gnome.org/browse/banshee/commit/?id=f847cb0c742b2d993dad71dba373bd81bb457afa

+++ src/Core/Banshee.Services/Banshee.Collection.Database/DatabaseImportManager.cs
@@ -34,13 +34,10 @@ using System.IO;
 using Mono.Unix;
 
 using Hyena;
-using Hyena.Data.Sqlite;
 
-using Banshee.Base;
 using Banshee.Sources;
 using Banshee.Collection;
 using Banshee.Collection.Database;
-using Banshee.Configuration.Schema;
 using Banshee.ServiceStack;
 using Banshee.Streaming;
 
@@ -251,7 +248,7 @@ namespace Banshee.Collection.Database
 
         public void NotifyAllSources ()
         {
-            foreach (int primary_source_id in counts.Keys) {
+            foreach (long primary_source_id in counts.Keys) {
                 PrimarySource.GetById (primary_source_id).NotifyTracksAdded ();
             }
             counts.Clear ();
